Feat: 이벤트 DTO 리팩터링
- 이벤트 요청 DTO 리팩터링
- 응답 DTO 변경
- 이벤트 인터페이스 추가
- 이벤트 서비스 리팩터링

Related to #26
