+++ src/event/dto/get-event-list.dto.ts
@@ -1,4 +1,4 @@
-import { Exclude, Expose, Type } from 'class-transformer';
+import { Exclude, Expose } from 'class-transformer';
 import { ApiProperty } from '@nestjs/swagger';
 import { PhotoBoothBrand } from '../../brand/entity/brand.entity';
 import { Events } from '../entity/event.entity';
@@ -65,7 +65,6 @@ export class GetEventListDto {
     description: '이벤트와 관련된 해시태그',
   })
   @Expose()
-  @Type(() => EventHashtag)
   get hashtags(): string[] {
     return this._eventHashtags.map((hashtags) => hashtags.hashtag.name);
   }
